ml/requirements.txt
torch>=2.3.0
pyro-ppl>=1.9.0
numpy>=1.26.0
pandas>=2.2.0
cdsapi>=0.7.0
netCDF4>=1.7.0
xarray>=2024.1.0
scipy>=1.12.0
matplotlib>=3.8.0
tqdm>=4.66.0
requests>=2.31.0
executorch>=0.2.0
